feat(profil): import de toutes les composantes sur la nouvelle page
