+++ packages/frontend/src/app/modules/usager-profil/components/profil-courriers/profil-courriers.component.html
@@ -5,6 +5,12 @@
     <div class="page-content px-4 pb-5 mb-5">
       <div class="row">
         <div class="col-md-12">
+          <h4>Préférence de contact</h4>
+          <app-profil-edit-preference
+            [usager]="usager"
+          ></app-profil-edit-preference>
+        </div>
+        <div class="col-md-12">
           <h4>Transfert</h4>
           <app-profil-transfert-courrier
             [usager]="usager"
