+++ packages/frontend/src/app/modules/usager-profil/components/profil-dossier/profil-dossier.component.html
@@ -51,6 +51,18 @@
             </button>
           </div>
         </div>
+
+        <div class="col-md-12">
+          <app-documents [usager]="usager"></app-documents>
+
+          <app-upload
+            *ngIf="usager.isActif || usager.decision.statut === 'RADIE'"
+            [usager]="usager"
+          >
+          </app-upload>
+        </div>
+
+        <br />
       </div>
     </div>
   </div>
