+++ packages/frontend/src/app/modules/usager-profil/components/profil-head/profil-head.component.ts
@@ -1,8 +1,14 @@
-import { Component, Input, OnInit } from "@angular/core";
+import {
+  Component,
+  Input,
+  OnInit,
+  TemplateRef,
+  ViewChild,
+} from "@angular/core";
 import { ActivatedRoute, Router } from "@angular/router";
 import { NgbModal } from "@ng-bootstrap/ng-bootstrap";
 import { ToastrService } from "ngx-toastr";
-import { AppUser, UserRole } from "../../../../../_common/model";
+import { AppUser, UsagerLight, UserRole } from "../../../../../_common/model";
 import { AuthService } from "../../../shared/services/auth.service";
 import { NgbDateCustomParserFormatter } from "../../../shared/services/date-formatter";
 import { UsagerFormModel } from "../../../usagers/components/form/UsagerFormModel";
@@ -19,6 +25,10 @@ export class ProfilHeadComponent implements OnInit {
   @Input() public me: AppUser;
 
   public today: Date;
+
+  @ViewChild("renewModal", { static: true })
+  public renewModal!: TemplateRef<any>;
+
   constructor(
     private interactionService: InteractionService,
     private authService: AuthService,
@@ -33,11 +43,32 @@ export class ProfilHeadComponent implements OnInit {
   }
 
   public ngOnInit(): void {}
+
   public closeModals() {
     this.modalService.dismissAll();
   }
 
   public isRole(role: UserRole) {
     return this.me.role === role;
+  }
+
+  public renouvellement() {
+    this.usagerService.renouvellement(this.usager.ref).subscribe(
+      (usager: UsagerLight) => {
+        this.usager = new UsagerFormModel(usager);
+        this.modalService.dismissAll();
+        this.router.navigate(["usager/" + usager.ref + "/edit"]);
+        this.notifService.success(
+          "Votre demande a été enregistrée. Merci de remplir l'ensemble du dossier"
+        );
+      },
+      (error) => {
+        this.notifService.error("Impossible d'enregistrer cette interaction");
+      }
+    );
+  }
+
+  public openRenewModal() {
+    this.modalService.open(this.renewModal);
   }
 }
