+++ packages/frontend/src/app/modules/usager-profil/components/profil-historique-courriers/profil-historique-courriers.component.ts
@@ -36,7 +36,10 @@ export class ProfilHistoriqueCourriersComponent implements OnInit {
     private matomo: MatomoTracker,
     private interactionService: InteractionService
   ) {}
-  public ngOnInit(): void {}
+  public ngOnInit(): void {
+    console.log("ook");
+    this.getInteractions();
+  }
 
   public deleteInteraction(idInteraction: number) {
     this.matomo.trackEvent("profil", "interactions", "delete", 1);
