+++ packages/frontend/src/app/modules/usager-profil/components/profil-overview/profil-overview.component.html
@@ -9,7 +9,7 @@
             <h4 class="col-12">Dossier</h4>
             <div class="col-6"><span class="label">Statut</span></div>
             <div class="col-6">
-              <span class="label-info statut-label">
+              <span class="statut-label">
                 <span [class]="usager.statutColor">
                   {{ DECISION_STATUT_LABELS[usager.decision.statut] }}
                 </span>
@@ -25,15 +25,56 @@
             </div>
           </div>
         </div>
-        <div class="col-md-6"><h4>Courrier en attente</h4></div>
+
+        <div class="col-md-6">
+          <h4>Courrier en attente</h4>
+          <div class="profil-interaction my-3">
+            <div class="interaction-icon-box mr-3">
+              <div class="interaction-icon icon-courrier"></div>
+            </div>
+            <div class="profil-interaction-label">
+              <b>{{ usager.lastInteraction.courrierIn }}</b> courriers
+            </div>
+          </div>
+          <div class="profil-interaction my-3">
+            <div class="interaction-icon-box mr-3">
+              <div class="interaction-icon icon-colis"></div>
+            </div>
+            <div class="profil-interaction-label">
+              <b>{{ usager.lastInteraction.colisIn }}</b> colis
+            </div>
+          </div>
+          <div class="profil-interaction my-3">
+            <div class="interaction-icon-box mr-3">
+              <div class="interaction-icon icon-recommande"></div>
+            </div>
+            <div class="profil-interaction-label">
+              <b>{{ usager.lastInteraction.recommandeIn }}</b> avis de passage
+            </div>
+          </div>
+        </div>
+
         <div class="col-md-6">
           <h4>Documents téléchargeables</h4>
 
+          <button (click)="getAttestation()" class="btn btn-primary">
+            <fa-icon icon="download" class="mr-2"></fa-icon>
+            Télécharger le CERFA
+          </button>
+          <br />
+          <br />
           <app-profil-structure-docs
             [usager]="usager"
           ></app-profil-structure-docs>
         </div>
-        <div class="col-md-6"><h4>Historique des interactions</h4></div>
+        <div class="col-md-6">
+          <h4>Historique des interactions</h4>
+
+          <app-profil-historique-courriers
+            [usager]="usager"
+            [me]="me"
+          ></app-profil-historique-courriers>
+        </div>
       </div>
     </div>
   </div>
