--- packages/frontend/src/app/modules/usagers/components/profil/profil.css
@@ -1,222 +1,0 @@
-#notif-container {
-  max-width: 800px;
-  margin: 0 auto;
-}
-
-#notif_button,
-#notif-form,
-#big_icons {
-  display: inline-block;
-  vertical-align: middle;
-}
-
-#notif-form {
-  text-align: left;
-  width: 290px;
-}
-#notif_button {
-  width: 120px;
-}
-#notif-form div {
-  margin: 5px auto;
-}
-
-input[type="number"] {
-  width: 60px;
-  margin: 0 10px;
-  display: inline-block;
-}
-
-.actions_icones:hover {
-  color: var(--secondary);
-}
-
-.actions_icones {
-  display: block;
-  color: white;
-  -webkit-text-stroke-width: 12px;
-  stroke-width: 10px;
-  transition: all 0.3s;
-  cursor: pointer;
-  font-size: 1.3rem;
-  padding: 10px;
-}
-
-#big_icons div {
-  border-radius: 5px;
-  padding: 0 5px;
-  text-align: center;
-  display: inline-block;
-  vertical-align: middle;
-}
-
-#big_icons {
-  margin-left: 30px;
-  width: 340px;
-  border-left: 1px solid #dfdfdf;
-  text-align: center;
-}
-
-#big_icons span {
-  font-size: 0.8rem;
-  line-height: 1rem;
-  padding-top: 5px;
-  padding-bottom: 20px;
-  display: block;
-}
-
-.btn-pink {
-  background-color: rgba(189, 16, 224, 0.5);
-  color: white;
-}
-
-.btn-pink:hover {
-  background-color: #bd10e0;
-}
-
-.interactions-actions {
-  width: 50px;
-  padding: 0.3rem;
-}
-
-.delete-interaction .btn {
-  margin: 0 5px;
-}
-
-.delete-interaction {
-  padding: 0.5rem;
-  text-align: center;
-  font-weight: 600;
-  background: rgba(247, 247, 247, 0.95);
-}
-
-.interactions-actions .edit:hover {
-  background: var(--primary);
-}
-.interactions-actions .delete:hover {
-  background: var(--danger);
-}
-
-.interactions-actions .ng-fa-icon:hover {
-  color: white;
-}
-
-.interactions-actions .ng-fa-icon {
-  padding: 6px 5px;
-  border-radius: 6px;
-  transition: all 0.2s;
-  width: 35px;
-  text-align: center;
-  cursor: pointer;
-  display: inline-block;
-}
-
-#interactions-table thead,
-#interactions-table tbody tr {
-  display: table;
-  width: 99%;
-  table-layout: fixed; /* even columns width , fix width of table too*/
-}
-#interactions-table > tbody > tr:hover > td.interactions-actions > span {
-  visibility: visible;
-  opacity: 1;
-}
-
-#interactions-table > tbody > tr > td.interactions-actions > span {
-  visibility: hidden;
-  opacity: 0;
-}
-
-#interactions-table tbody {
-  display: block;
-  height: 250px;
-  overflow: auto;
-}
-
-#delete-usager:hover {
-  cursor: pointer;
-  text-decoration: underline;
-}
-
-#ayants-droit-table th,
-#ayants-droit-table td {
-  width: 33.3%;
-}
-
-.profil-table th:first-child,
-.profil-table td:first-child {
-  width: 25%;
-}
-
-.profil-table th:nth-child(2),
-.profil-table td:nth-child(2) {
-  width: 45%;
-}
-.profil-table th:nth-child(3),
-.profil-table td:nth-child(3) {
-  width: 20%;
-}
-
-.profil-table th:nth-child(4),
-.profil-table td:nth-child(4) {
-  width: 5%;
-}
-
-.sous-titre {
-  color: #212529;
-  text-transform: uppercase;
-  display: block;
-  margin: 10px 0;
-  font-size: 1.02rem;
-}
-
-.text-radiation {
-  font-size: 1.15rem;
-  font-weight: 600;
-}
-
-.modal-title {
-  font-size: 1.3rem;
-  font-weight: 600;
-}
-
-/* OLD TO ERASE */
-
-.appel {
-  -webkit-text-stroke-color: var(--secondary);
-  stroke: var(--secondary);
-}
-
-.appel:hover {
-  color: var(--secondary) !important;
-}
-
-.visite {
-  -webkit-text-stroke-color: var(--primary);
-  stroke: var(--primary);
-}
-
-.visite:hover {
-  color: #4668ff !important;
-}
-
-.bannir {
-  -webkit-text-stroke-color: var(--danger);
-  stroke: var(--danger);
-}
-
-.bannir:hover {
-  color: var(--danger) !important;
-}
-
-.npai {
-  -webkit-text-stroke-width: 0;
-  stroke-width: 0;
-  color: var(--danger) !important;
-}
-
-.courrier {
-  -webkit-text-stroke-color: #c62bd9;
-  stroke: #c62bd9;
-  color: white;
-}
